hw_v2

+++ log_parser.py
@@ -2,54 +2,18 @@ import json
 import os
 import argparse
 import re
-from collections import defaultdict
 
 parser = argparse.ArgumentParser(description="путь для директории")
 parser.add_argument("-f", dest="path", default="logs/", action='store', help="Path to logfile")
-
 args = parser.parse_args()
 
-def files_in_derectory(path):
-    j = 1
-    file_list_of_directory = os.listdir(path)
-    for i in file_list_of_directory:
-        print(j, ". ", i)
-        j += 1
-    try:
-        choose_number = int(input('введите номер файла для анализа логов:'))
-    except:
-        "нет такого номера"
-    try:
-        choose_file = file_list_of_directory[choose_number - 1]
-        print(f"выбран файл {choose_file}, обработка может занять некоторое время")
-    except:
-        choose_file = None
-        print("вы не выбрали файл или нет файла с таким номером")
-        exit()
-    return choose_file
 
-def log_parser(file: str):
-    dict_ip = defaultdict(
-        lambda: {'GET': 0, 'POST': 0, 'PUT': 0, 'PATCH': 0, 'DELETE': 0, 'HEAD': 0}
-    )
-    with open(args.path + f'/{file}') as log:
-        for line in log:
-            ip_match = re.search(r"\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}", line)
-            if ip_match is not None:
-                ip = ip_match.group()
-                method = re.search(r'\] \"(POST|GET|PUT|PATCH|DELETE|HEAD)', line)
-                if method is not None:
-                    method = method.groups()[0]
-                    dict_ip[ip][method] += 1
-        return dict_ip
-
-
-def top_3_log_parser(file: str):
+def top_3_log_parser(agr_path, file_of_log: str, kostil='/'):
     """топ 3 самых долгих запросов, url, ip, время запроса """
     top_requests = {'#1': {'request': '', 'ip': '', 'url': '', 'time': 0},
                     '#2': {'request': '', 'ip': '', 'url': '', 'time': 0},
                     '#3': {'request': '', 'ip': '', 'url': '', 'time': 0}}
-    with open(args.path + f'/{file}') as log:
+    with open(agr_path + kostil + file_of_log) as log:
         for line in log:
             ip_match = re.search(r"\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}", line)
             time_r = re.search(r'\d+$', line)
@@ -72,10 +36,10 @@ def top_3_log_parser(file: str):
         return top_requests
 
 
-def count_request_parser(file: str):
+def count_request_parser(agr_path, file_of_log: str, kostil='/'):
     """общее количество выполненных запросов"""
     count_request = 0
-    with open(args.path + f'/{file}') as log:
+    with open(agr_path + kostil + file_of_log) as log:
         for line in log:
             ip_match = re.search(r"\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}", line)
             if ip_match is not None:
@@ -85,13 +49,13 @@ def count_request_parser(file: str):
         return count_request
 
 
-def top_3_ip_cont_request(file: str):
+def top_3_ip_cont_request(agr_path, file_of_log: str, kostil='/'):
     '''топ 3 IP адресов, с которых были сделаны запросы'''
     cont_requests = {}
     top_ip = {"#1": {'ip': '', 'count': 0},
               "#2": {'ip': '', 'count': 0},
               "#3": {'ip': '', 'count': 0}}
-    with open(args.path + f'/{file}') as log:
+    with open(agr_path + kostil + file_of_log) as log:
         for line in log:
             ip_match = re.search(r"\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}", line)
             if ip_match is not None:
@@ -110,10 +74,10 @@ def top_3_ip_cont_request(file: str):
         return top_ip
 
 
-def counter_requests(file: str):
+def counter_requests(agr_path, file_of_log: str, kostil='/'):
     """количество запросов по типу: GET - 20, POST - 10 и т.п."""
     dict_requests = {'GET': 0, 'POST': 0, 'PUT': 0, 'PATCH': 0, 'DELETE': 0, 'HEAD': 0}
-    with open(args.path + f'/{file}') as log:
+    with open(agr_path + kostil + file_of_log) as log:
         for line in log:
             ip_match = re.search(r"\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}", line)
             if ip_match is not None:
@@ -125,17 +89,31 @@ def counter_requests(file: str):
               json.dumps(dict_requests, indent=4, sort_keys=True))
         return dict_requests
 
+
 try:
     os.mkdir("result")
 except:
     pass
 
-for file in os.listdir(args.path):
-    with open(f"result/log_parser--{file}.json", 'w') as result_file:
-        json.dump(top_3_log_parser(file), result_file, indent=4)
+
+if os.path.isfile(args.path):
+    file = ""
+    name = args.path.split("/")[:0:-1]
+    with open(f"result/log_parser--{name[0]}.json", 'w') as result_file:
+        json.dump(top_3_log_parser(args.path, file, kostil=''), result_file, indent=4)
         result_file.write(",\r\n")
-        json.dump(counter_requests(file), result_file, indent=4)
+        json.dump(counter_requests(args.path, file, kostil=''), result_file, indent=4)
         result_file.write(",\r\n")
-        json.dump(top_3_ip_cont_request(file), result_file, indent=4)
+        json.dump(top_3_ip_cont_request(args.path, file, kostil=''), result_file, indent=4)
         result_file.write(",\r\n")
-        json.dump(count_request_parser(file), result_file, indent=2)
+        json.dump(count_request_parser(args.path, file, kostil=''), result_file, indent=2)
+else:
+    for file in os.listdir(args.path):
+        with open(f"result/log_parser--{file}.json", 'w') as result_file:
+            json.dump(top_3_log_parser(args.path, file), result_file, indent=4)
+            result_file.write(",\r\n")
+            json.dump(counter_requests(args.path, file), result_file, indent=4)
+            result_file.write(",\r\n")
+            json.dump(top_3_ip_cont_request(args.path, file), result_file, indent=4)
+            result_file.write(",\r\n")
+            json.dump(count_request_parser(args.path, file), result_file, indent=2)
